[#111] refactor: jwt 관련 파일 이동
